+++ src/components/PullComponents.vue
@@ -45,6 +45,14 @@ export default {
       type: Number,
       default: 0,
     },
+    // 返回的数据是对象还是数组
+    isObj: {
+      type: Object,
+      default: () => ({
+        type: 'Array',
+        key: '',
+      }),
+    },
   },
   computed: {
     displayAboutScroll() {
@@ -66,20 +74,28 @@ export default {
       const params = this.params || {};
       params.page = this.page;
       axios.get(`${apiServer}/${this.apiUrl}`, { params }).then(({ data }) => {
-        // 清空数组 ps: 如果在 refresh 里面清空数组
-        // 1.点击的时会先执行触摸刷新的方法 导致无法正常单击切换页面
-        // 2.因为先执行触摸方法 清空了数组 会给页面造成影响
         if (isEmptyArray) this.articles.length = 0;
-        // Merge arrays with destruction
-        this.articles = [...this.articles, ...data];
-        this.$emit('getListData', {
-          data: this.articles,
-          index: this.nowIndex,
-        });
-        if (data.length >= 0 && data.length < 20) this.isTheEndOfTheScroll = true;
-        else this.page += 1;
+        if (this.isObj.type === 'Array') {
+          // 如果返回的数据是 Array 返回整个 data
+          this.articles = [...this.articles, ...data];
+          this.$emit('getListData', {
+            data: this.articles,
+            index: this.nowIndex,
+          });
+          if (data.length >= 0 && data.length < 20) this.isTheEndOfTheScroll = true;
+        } else if (this.isObj.type === 'Object') {
+          // 如果返回的是 Object 根据传进来的字段获取相应的 list
+          const resData = data[this.isObj.key];
+          this.articles = [...this.articles, ...resData];
+          this.$emit('getListData', {
+            data,
+            list: this.articles,
+            index: this.nowIndex,
+          });
+          if (resData.length >= 0 && resData.length < 20) this.isTheEndOfTheScroll = true;
+        }
+        this.page += 1;
         this.busy = false;
-        // 列表最后一列小于二十显示加载完
       }).catch((err) => {
         console.log(err);
         this.$Message.error('获取文章发生错误');
@@ -112,8 +128,9 @@ export default {
 <style scoped>
 /* 加载更多提示 */
 .loading-stat {
-  margin: 10px;
+  margin: 20px 0;
   color: #999;
-  font-size: 13px;
+  font-size: 14px;
+  text-align: center;
 }
 </style>
