+++ src/views/User/Asset/Asset.less
@@ -1,0 +1,99 @@
+.assetpage {
+  background:rgba(247,247,247,1);
+  opacity:1;
+  text-align: left;
+}
+.my-assets {
+  margin: auto;
+  margin-top: 13px;
+  text-align: center;
+  max-height: 128px;
+  /* margin: -32px 20px 0 20px; */
+  /*padding: 8px;*/
+  background: rgba(255, 255, 255, 1);
+  box-shadow: 0px 5px 5px 0px rgba(213, 213, 213, 0.5);
+  border-radius: 4px;
+}
+.row1 {
+  font: 12px;
+  /*text-align: left;*/
+}
+.row2 {
+  text-align: center;
+}
+.title-bar {
+  white-space: nowrap;
+  font-size: 12px;
+  font-family: PingFangSC-Semibold;
+  line-height: 0px;
+  color: rgba(152,152,152,1);
+  opacity: 1;
+}
+.income {
+  color: #FF3030;
+  font-size: 16px;
+  font-family: PingFangSC-Semibold;
+}
+.income2 {
+  color:  #07BB3D;
+  font-size: 16px;
+  font-family: PingFangSC-Semibold;
+}
+.assets {
+  /* background: rgba(240, 240, 240, 1); */
+  margin-top: -100px;
+  padding-top: 100px;
+  font-weight: bold;
+}
+.topcard{
+  background-color: #ffffff;
+  margin: 20px;
+  padding: 20px;
+  padding-bottom: 15px;
+  background: rgba(255, 255, 255, 1);
+  box-shadow: 0px 2px 5px 3px rgba(233, 233, 233, 0.5);
+  border-radius: 8px;
+}
+.toptext1{
+  color: rgba(0, 0, 0, 1);
+  font-size: 12px;
+  opacity: 0.4;
+  float: left;
+}
+.topremain{
+  color: rgba(0, 0, 0, 1);
+  font-size: 22px;
+  font-weight: bolder;
+  float: top;
+}
+.toptext2{
+  text-align: center;
+  color: rgba(0, 0, 0, 1);
+  font-size: 12px;
+}
+.toptext3{
+  padding-top: 7px;
+  text-align: center;
+  color: rgba(0, 0, 0, 1);
+  font-size: 14px;
+  font-weight: bold;
+}
+.detailtext{
+  font-size: 13px;
+  margin-left: 20px;
+  font-weight: bold;
+}
+Button.withdraw, Button.withdraw:focus, Button.withdraw:hover {
+  background-color: rgba(0, 0, 0, 1);
+  border-radius: 2px;
+  color: rgba(255,255,255,1);
+  float: right;
+  font-size: 12px;
+  margin-top: 20px;
+  width: 80px;
+  height: 25px;
+  letter-spacing: 2px;
+  max-width: 94px;
+  max-height: 35px;
+  margin-right: 0px;
+}
