+++ src/views/User/DraftBox.vue
@@ -67,7 +67,7 @@ export default {
   },
 };
 </script>
-<style>
+<style scoped>
 a {
   color: #000;
   text-decoration: none; /* no underline */
